refactor: enhance skills section with dark mode
