--- src/pages/Home/skills.css
@@ -1,128 +1,0 @@
-/* ------------------------------------------- */
-/*            SIDEBAR - MODO CLARO            */
-/* ------------------------------------------- */
-
-.sidebar {
-  border-radius: 16px;
-  background: linear-gradient(135deg, #f9f9f9, #eaeaea);
-  box-shadow: -6px -6px 12px rgba(255, 255, 255, 0.8),
-              6px 6px 12px rgba(0, 0, 0, 0.1);
-  transition: all 0.4s ease;
-}
-
-.sidebar:hover {
-  transform: scale(1.02);
-}
-
-/* ------------------------------------------- */
-/*            SIDEBAR - MODO OSCURO           */
-/* ------------------------------------------- */
-
-.dark .sidebar {
-  background: linear-gradient(135deg, #1c1c1c, #2b2b2b);
-  box-shadow: -4px -4px 8px rgba(255, 255, 255, 0.03),
-              4px 4px 8px rgba(0, 0, 0, 0.2);
-  transition: all 0.4s ease;
-}
-
-.dark .sidebar:hover {
-  transform: scale(1.02);
-}
-
-/* ------------------------------------------- */
-/*          CARD - MODO CLARO                 */
-/* ------------------------------------------- */
-
-.card {
-  border-radius: 12px;
-  background: linear-gradient(145deg, #ffffff, #e6e6e6);
-  box-shadow: -4px -4px 8px rgba(255, 255, 255, 0.8),
-              4px 4px 8px rgba(0, 0, 0, 0.1);
-  transition: transform 0.4s ease, box-shadow 0.4s ease;
-}
-
-.card:hover {
-  transform: translateY(-8px) scale(1.05);
-  box-shadow: -6px -6px 12px rgba(255, 255, 255, 0.9),
-              6px 6px 12px rgba(0, 0, 0, 0.15);
-}
-
-/* ------------------------------------------- */
-/*          CARD - MODO OSCURO                */
-/* ------------------------------------------- */
-
-.dark .card {
-  background: linear-gradient(145deg, #252525, #313131);
-  box-shadow: -4px -4px 8px rgba(255, 255, 255, 0.03),
-              4px 4px 8px rgba(0, 0, 0, 0.2);
-}
-
-.dark .card:hover {
-  transform: translateY(-8px) scale(1.05);
-  box-shadow: -6px -6px 12px rgba(255, 255, 255, 0.05),
-              6px 6px 12px rgba(0, 0, 0, 0.25);
-}
-
-.dark .card h3 {
-  color: rgba(240, 240, 240, 0.85);
-  transition: color 0.4s ease;
-}
-
-.dark .card:hover h3 {
-  color: rgba(255, 255, 255, 1);
-}
-
-.dark .card .icon {
-  color: rgba(200, 200, 200, 0.8);
-  transition: color 0.4s ease, transform 0.4s ease;
-}
-
-.dark .card:hover .icon {
-  color: rgba(255, 255, 255, 0.9);
-  transform: scale(1.2);
-}
-
-/* ------------------------------------------- */
-/*          ANIMACIÓN DE BLOB                */
-/* ------------------------------------------- */
-
-@keyframes blob {
-  0% {
-    transform: translate(0, 0) scale(1);
-  }
-  33% {
-    transform: translate(30px, -50px) scale(1.2);
-  }
-  66% {
-    transform: translate(-20px, 20px) scale(0.9);
-  }
-  100% {
-    transform: translate(0, 0) scale(1);
-  }
-}
-
-.animate-blob {
-  animation: blob 10s infinite;
-  animation-timing-function: ease-in-out;
-  animation-fill-mode: forwards;
-}
-
-.animation-delay-2000 {
-  animation-delay: 2s;
-}
-
-.animation-delay-4000 {
-  animation-delay: 4s;
-}
-
-.blob {
-  pointer-events: none;
-  z-index: 0;
-}
-
-.dark .blob {
-  background: rgba(150, 150, 255, 0.2);
-  mix-blend-mode: screen;
-  filter: blur(90px);
-  opacity: 0.85;
-}
